Add creator validation

+++ src/controllers/movieController.js
@@ -64,7 +64,11 @@ movieController.post('/:movieId/attach', async (req, res) => {
 movieController.get('/:movieId/delete', isAuth, async (req, res) => {
     const movieId = req.params.movieId;
     
-    // TODO Check if creator
+    const movie = await movieService.getOne(movieId);
+    
+    if (!movie.creator?.equals(req.user.id)) {
+        return res.redirect('/');
+    }
 
     await movieService.delete(movieId);
